pyproject.toml
[project]
name = "attachiq"
version = "0.1.0"
description = "AttachIQ: Multimodal Request-Attachment Triage for AI Assistants"
readme = "README.md"
requires-python = ">=3.13,<3.14"
authors = [
    { name = "Mohammad Abu Jafar" },
]
dependencies = [
    "torch>=2.9.0,<2.10",
    "torchvision>=0.24.0,<0.25",
    "transformers>=4.46.3",
    "safetensors>=0.4.5",
    "huggingface-hub>=0.26.5",
    "datasets>=3.1.0",
    "numpy>=2.1,<3",
    "pandas>=2.2.3",
    "scikit-learn>=1.5.2",
    "pydantic>=2.9,<3",
    "loguru>=0.7.2",
    "matplotlib>=3.9.2",
    "pillow>=11.0.0",
    "tqdm>=4.67.1",
    "streamlit>=1.40.0",
    "typer>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "ruff==0.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/attachiq"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.ruff]
line-length = 110
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
